desktop/requirements.txt
PySide6==6.7.2
pyserial==3.5
requests==2.32.3
